Introduce `ChunkedLoggingStreamTestUtils`
Today the test utilities for extracting data logged by a
`ChunkedLoggingStream` are in the `:server` test suite which renders
them inaccessible to the test suites of other modules. This commit moves
them to `:test:framework`.
